// src/models/transferencia-por-moler.model.ts
export class DataPorMolerModel {
    login: string;
    data:  Data2;
}

export class Data2 {
    produccionOF:        string;
    bodega:              string;
    cantidadCajas:       number;
    fecha:               Date;
    lote:                string;
    cantidad:            number;
    total:               string;
    precioEnvase:        string;
    descripcionProducto: string;
    codProceso:          string;
    proceso:             string;
    cantxCaja:           string;
    color:               string;
    tipoResina:          string;
    codCosto:            string;
    familia:             string;
    peso:                string;
    codigoPorMoler:      string;
    codigoProducto:      string;
    cliente:             string;
    nombreMaquina:       string;
    precio:              string;
    unidadMedida:        string;
    ipImpresora:         string;
}